#1766 move the commonly used dialog to core-ui
