+++ core/ui/src/main/resources/org/phoebus/ui/messages.properties
@@ -1,5 +1,11 @@
+Add=Add
 AddColumn=Add Column
 AddRow=Add Row
+Add_Tooltip=Add the selected items.
+Apply=Apply
+Available=Available
+Clear=Clear
+Clear_Tooltip=Clear the selected items list.
 DefaultNewColumnName=New Column
 DoNotShow=Don't show this message again.
 Format_Binary=Binary
@@ -20,6 +26,7 @@ MoveColumnRight=Move Column Right
 MoveRowDown=Move Row Down
 MoveRowUp=Move Row Up
 NumberInputHdr=Please enter a number
+Num_Selected=selected
 Redo_TT=Re-do last change
 RemoveColumn=Remove Column
 RemoveRow=Remove Row
@@ -27,4 +34,7 @@ RenameColumn=Rename Column
 RenameColumnInfo=Please enter desired column name
 RenameColumnTitle=Column Name
 ReplaceExamplesWarningFMT=The directory\n  {0}\nis not empty.\n\nContinue to replace the existing content of\nthe directory with the latest examples?
+Search=Search available:
+SearchAvailableItems=Search the list of available items.
+Selected=Selected
 Undo_TT=Un-do last change
